Refactor invoice schema to separate file
Moved getInvoiceSchema and GetInvoiceResponse type from zuoraSchemas.ts to a new dedicated invoice.ts file for better modularity and organization.
